refactor: standalone d3d and ui
