+++ src/Util.h
@@ -1,34 +1,7 @@
 #pragma once
 
+#include "Utils/D3D.h"
 #include "Utils/Format.h"
 #include "Utils/Game.h"
 #include "Utils/Serialize.h"
-
-namespace Util
-{
-
-	ID3D11ShaderResourceView* GetSRVFromRTV(ID3D11RenderTargetView* a_rtv);
-	ID3D11RenderTargetView* GetRTVFromSRV(ID3D11ShaderResourceView* a_srv);
-	std::string GetNameFromSRV(ID3D11ShaderResourceView* a_srv);
-	std::string GetNameFromRTV(ID3D11RenderTargetView* a_rtv);
-	void SetResourceName(ID3D11DeviceChild* Resource, const char* Format, ...);
-	ID3D11DeviceChild* CompileShader(const wchar_t* FilePath, const std::vector<std::pair<const char*, const char*>>& Defines, const char* ProgramType, const char* Program = "main");
-
-	/**
-	 * Usage:
-	 * if (auto _tt = Util::HoverTooltipWrapper()){
-	 *     ImGui::Text("What the tooltip says.");
-	 * }
-	*/
-	class HoverTooltipWrapper
-	{
-	private:
-		bool hovered;
-
-	public:
-		HoverTooltipWrapper();
-		~HoverTooltipWrapper();
-		inline operator bool() { return hovered; }
-	};
-
-}
+#include "Utils/UI.h"
