+++ src/Utils/D3D.h
@@ -1,0 +1,12 @@
+#pragma once
+
+namespace Util
+{
+	ID3D11ShaderResourceView* GetSRVFromRTV(ID3D11RenderTargetView* a_rtv);
+	ID3D11RenderTargetView* GetRTVFromSRV(ID3D11ShaderResourceView* a_srv);
+	std::string GetNameFromSRV(ID3D11ShaderResourceView* a_srv);
+	std::string GetNameFromRTV(ID3D11RenderTargetView* a_rtv);
+	void SetResourceName(ID3D11DeviceChild* Resource, const char* Format, ...);
+
+	ID3D11DeviceChild* CompileShader(const wchar_t* FilePath, const std::vector<std::pair<const char*, const char*>>& Defines, const char* ProgramType, const char* Program = "main");
+}  // namespace Util
