+++ src/Utils/UI.cpp
@@ -1,0 +1,41 @@
+#include "UI.h"
+
+namespace Util
+{
+	HoverTooltipWrapper::HoverTooltipWrapper()
+	{
+		hovered = ImGui::IsItemHovered();
+		if (hovered) {
+			ImGui::BeginTooltip();
+			ImGui::PushTextWrapPos(ImGui::GetFontSize() * 35.0f);
+		}
+	}
+
+	HoverTooltipWrapper::~HoverTooltipWrapper()
+	{
+		if (hovered) {
+			ImGui::PopTextWrapPos();
+			ImGui::EndTooltip();
+		}
+	}
+
+	DisableGuard::DisableGuard(bool disable) :
+		disable(disable)
+	{
+		if (disable)
+			ImGui::BeginDisabled();
+	}
+	DisableGuard::~DisableGuard()
+	{
+		if (disable)
+			ImGui::EndDisabled();
+	}
+
+	bool PercentageSlider(const char* label, float* data, float lb, float ub, const char* format)
+	{
+		float percentageData = (*data) * 1e2f;
+		bool retval = ImGui::SliderFloat(label, &percentageData, lb, ub, format);
+		(*data) = percentageData * 1e-2f;
+		return retval;
+	}
+}  // namespace Util
